B07-880 branch C unnecessary rescaling of x axis
-Change spectrum data type to double
-Create a copy of spectrum in each datapoint and use for event system
-Call get spectrum with a fixed size parameter
-Cache current photon energy at new region and reuse

+++ uk.ac.diamond.daq.devices.specs.phoibos.api/src/uk/ac/diamond/daq/devices/specs/phoibos/api/SpecsPhoibosLiveDataUpdate.java
@@ -112,7 +112,7 @@ public class SpecsPhoibosLiveDataUpdate extends SpecsPhoibosLiveUpdate {
 		private int currentPoint;
 		private int totalIterations;
 		private int currentPointInIteration;
-		private float[] spectrum;
+		private double[] spectrum;
 		private double[][] image;
 		private double[] keEnergyAxis;
 		private double[] beEnergyAxis;
@@ -149,7 +149,7 @@ public class SpecsPhoibosLiveDataUpdate extends SpecsPhoibosLiveUpdate {
 			return this;
 		}
 
-		public Builder spectrum(float[] val) {
+		public Builder spectrum(double[] val) {
 			spectrum = val;
 			return this;
 		}
